+++ async_rithmic/enums.py
@@ -20,18 +20,3 @@ TimeBarType = pb.request_time_bar_replay_pb2.RequestTimeBarReplay.BarType
 InstrumentType = pb.request_search_symbols_pb2.RequestSearchSymbols.InstrumentType
 SearchPattern = pb.request_search_symbols_pb2.RequestSearchSymbols.Pattern
 
-class Gateway(str, enum.Enum):
-    TEST = "rituz00100.rithmic.com:443"
-
-    CHICAGO = "rprotocol.rithmic.com:443"
-    SYDNEY = "au.rithmic.com:443"
-    SAO_PAULO = "br.rithmic.com:443"
-    COLO75 = "colo75.rithmic.com:443"
-    FRANKFURT = "de.rithmic.com:443"
-    HONGKONG = "hk.rithmic.com:443"
-    IRELAND = "ie.rithmic.com:443"
-    MUMBAI = "in.rithmic.com:443"
-    SEOUL = "kr.rithmic.com:443"
-    CAPETOWN = "za.rithmic.com:443"
-    TOKYO = "jp.rithmic.com:443"
-    SINGAPORE = "sg.rithmic.com:443"